fix(reviews): added utils and updated tests
