+++ cpp/src/arrow/engine/substrait/relation_internal.cc
@@ -47,6 +47,7 @@
 #include "arrow/engine/substrait/relation.h"
 #include "arrow/engine/substrait/type_internal.h"
 #include "arrow/filesystem/filesystem.h"
+#include "arrow/engine/substrait/util.h"
 #include "arrow/filesystem/localfs.h"
 #include "arrow/filesystem/type_fwd.h"
 #include "arrow/filesystem/util_internal.h"
@@ -654,17 +655,29 @@ Result<DeclarationInfo> FromProto(const substrait::Rel& rel, const ExtensionSet&
       // Note: at the moment Acero only supports UNION_ALL operation
       switch (op) {
         case substrait::SetRel::SET_OP_UNSPECIFIED:
-          return Status::NotImplemented("NotImplemented union type");
+          return Status::NotImplemented(
+              "NotImplemented union type : ",
+              EnumToString(op, substrait::SetRel_SetOp_descriptor()));
         case substrait::SetRel::SET_OP_MINUS_PRIMARY:
-          return Status::NotImplemented("NotImplemented union type");
+          return Status::NotImplemented(
+              "NotImplemented union type : ",
+              EnumToString(op, substrait::SetRel_SetOp_descriptor()));
         case substrait::SetRel::SET_OP_MINUS_MULTISET:
-          return Status::NotImplemented("NotImplemented union type");
+          return Status::NotImplemented(
+              "NotImplemented union type : ",
+              EnumToString(op, substrait::SetRel_SetOp_descriptor()));
         case substrait::SetRel::SET_OP_INTERSECTION_PRIMARY:
-          return Status::NotImplemented("NotImplemented union type");
+          return Status::NotImplemented(
+              "NotImplemented union type : ",
+              EnumToString(op, substrait::SetRel_SetOp_descriptor()));
         case substrait::SetRel::SET_OP_INTERSECTION_MULTISET:
-          return Status::NotImplemented("NotImplemented union type");
+          return Status::NotImplemented(
+              "NotImplemented union type : ",
+              EnumToString(op, substrait::SetRel_SetOp_descriptor()));
         case substrait::SetRel::SET_OP_UNION_DISTINCT:
-          return Status::NotImplemented("NotImplemented union type");
+          return Status::NotImplemented(
+              "NotImplemented union type : ",
+              EnumToString(op, substrait::SetRel_SetOp_descriptor()));
         case substrait::SetRel::SET_OP_UNION_ALL:
           break;
         default:
