+++ cpp/src/arrow/engine/substrait/util.h
@@ -30,6 +30,8 @@
 #include "arrow/type_fwd.h"
 #include "arrow/util/macros.h"
 
+#include "substrait/algebra.pb.h"  // IWYU pragma: export
+
 namespace arrow {
 
 namespace engine {
@@ -52,6 +54,9 @@ ARROW_ENGINE_EXPORT Result<std::shared_ptr<Buffer>> SerializeJsonPlan(
 ARROW_ENGINE_EXPORT std::shared_ptr<ExtensionIdRegistry> MakeExtensionIdRegistry();
 
 ARROW_ENGINE_EXPORT const std::string& default_extension_types_uri();
+
+ARROW_ENGINE_EXPORT
+std::string EnumToString(int value, const google::protobuf::EnumDescriptor* descriptor);
 
 }  // namespace engine
